+++ packages/atomic/src/components/search/atomic-search-box/atomic-search-box.tsx
@@ -1,4 +1,14 @@
-import {Component, h, State, Prop, Listen, Watch, Element} from '@stencil/core';
+import {
+  Component,
+  h,
+  State,
+  Prop,
+  Listen,
+  Watch,
+  Element,
+  Event,
+  EventEmitter,
+} from '@stencil/core';
 import {
   SearchBox,
   SearchBoxState,
@@ -38,6 +48,7 @@ import {
 } from './search-suggestion';
 
 import {hasKeyboard, isMacOS} from '../../../utils/device-utils';
+import {RedirectionPayload} from './redirection-payload';
 
 /**
  * The `atomic-search-box` component creates a search box with built-in support for suggestions.
@@ -90,7 +101,7 @@ export class AtomicSearchBox {
   private leftPanelRef: HTMLElement | undefined;
   private rightPanelRef: HTMLElement | undefined;
   private querySetActions!: QuerySetActionCreators;
-  private pendingSuggestionEvents: SearchBoxSuggestionsEvent[] = [];
+  private suggestionEvents: SearchBoxSuggestionsEvent[] = [];
   private suggestions: SearchBoxSuggestions[] = [];
   @Element() private host!: HTMLElement;
 
@@ -144,6 +155,26 @@ export class AtomicSearchBox {
    */
   @Prop({reflect: true}) public clearFilters = true;
 
+  /**
+   * Event that is emitted when a standalone search box redirection is triggered. By default, the search box will directly change the URL and redirect accordingly, so if you want to handle the redirection differently, use this event.
+   *
+   * @example
+   * ```html
+   * <script>
+   *   document.querySelector('atomic-search-box').addEventListener((e) => {
+   *     e.preventDefault();
+   *     // handle redirection
+   *   });
+   * </script>
+   * ...
+   * <atomic-search-box redirection-url="/search"></atomic-search-box>
+   * ```
+   */
+  @Event({
+    eventName: 'redirect',
+  })
+  public redirect!: EventEmitter<RedirectionPayload>;
+
   @AriaLiveRegion('search-box')
   protected searchBoxAriaMessage!: string;
 
@@ -178,12 +209,9 @@ export class AtomicSearchBox {
           options: searchBoxOptions,
         });
 
-    this.suggestions.push(
-      ...this.pendingSuggestionEvents.map((event) =>
-        event(this.suggestionBindings)
-      )
+    this.suggestions = this.suggestionEvents.map((event) =>
+      event(this.suggestionBindings)
     );
-    this.pendingSuggestionEvents = [];
   }
 
   public componentWillUpdate() {
@@ -204,18 +232,20 @@ export class AtomicSearchBox {
     storage.setJSON(StorageItems.STANDALONE_SEARCH_BOX_DATA, data);
 
     this.searchBox.afterRedirection();
-    window.location.href = redirectTo;
+    const event = this.redirect.emit({redirectTo, value});
+    if (!event.defaultPrevented) {
+      window.location.href = redirectTo;
+    }
   }
 
   @Listen('atomic/searchBoxSuggestion/register')
   public registerSuggestions(event: CustomEvent<SearchBoxSuggestionsEvent>) {
     event.preventDefault();
     event.stopPropagation();
+    this.suggestionEvents.push(event.detail);
     if (this.searchBox) {
       this.suggestions.push(event.detail(this.suggestionBindings));
-      return;
     }
-    this.pendingSuggestionEvents.push(event.detail);
   }
 
   @Watch('redirectionUrl')
@@ -752,9 +782,12 @@ export class AtomicSearchBox {
             onInput={(e) => this.onInput((e.target as HTMLInputElement).value)}
             onKeyDown={(e) => this.onKeyDown(e)}
             onClear={() => this.searchBox.clear()}
-            aria-controls={this.popupId}
-            role="combobox"
-            aria-activedescendant={this.activeDescendant}
+            popup={{
+              id: this.popupId,
+              activeDescendant: this.activeDescendant,
+              expanded: this.isExpanded,
+              hasSuggestions: this.hasSuggestions,
+            }}
           />
           {this.renderSuggestions()}
           <SubmitButton