+++ packages/samples/atomic-react/src/pages/FoldedResultListPage.tsx
@@ -39,60 +39,93 @@ export const FoldedResultListPage: FunctionComponent = () => {
 function MyTemplate(result: FoldedResult) {
   return (
     <>
+      <style>{`
+        .salesforce-badge::part(result-badge-element) {
+          background-color: #44a1da;
+          color: white;
+        }
+
+        .field {
+          display: inline-flex;
+          align-items: center;
+        }
+
+        .field-label {
+          font-weight: bold;
+          margin-right: 0.25rem;
+        }
+      `}</style>
       <AtomicResultSectionBadges>
-        <AtomicResultBadge field="ec_brand" />
+        <AtomicResultBadge label="Salesforce" class="salesforce-badge" />
+        {result.result.raw.language && (
+          <AtomicResultBadge icon="https://raw.githubusercontent.com/Rush/Font-Awesome-SVG-PNG/master/black/svg/language.svg">
+            <AtomicResultMultiValueText field="language" />
+          </AtomicResultBadge>
+        )}
       </AtomicResultSectionBadges>
       <AtomicResultSectionVisual>
-        <AtomicResultImage field="ec_images" />
+        <img loading="lazy" src="https://picsum.photos/350" />
       </AtomicResultSectionVisual>
       <AtomicResultSectionTitle>
         <AtomicResultLink />
       </AtomicResultSectionTitle>
       <AtomicResultSectionTitleMetadata>
-        <AtomicResultRating field="ec_rating" />
-        <AtomicResultPrintableUri maxNumberOfParts={3} />
+        <AtomicResultPrintableUri
+          maxNumberOfParts={3}
+        ></AtomicResultPrintableUri>
       </AtomicResultSectionTitleMetadata>
-      <AtomicResultSectionEmphasized>
-        <AtomicResultNumber field="ec_price">
-          <AtomicFormatCurrency currency="USD" />
-        </AtomicResultNumber>
-      </AtomicResultSectionEmphasized>
       <AtomicResultSectionExcerpt>
-        <AtomicResultText field="ec_shortdesc" />
+        <AtomicResultText field="excerpt" />
       </AtomicResultSectionExcerpt>
       <AtomicResultSectionChildren>
-        {result.children.map((child) => (
-          <p>
-            <a href={child.result.clickUri}>{child.result.title}</a>
-          </p>
-        ))}
+        {!!result.children.length && (
+          <div>
+            <b>This result has children:</b>
+            <ul>
+              {result.children.map((child, i) => (
+                <li>
+                  <b>Child {i + 1}: </b>
+                  <a href={child.result.clickUri}>{child.result.title}</a>
+                </li>
+              ))}
+            </ul>
+          </div>
+        )}
       </AtomicResultSectionChildren>
       <AtomicResultSectionBottomMetadata>
         <AtomicResultFieldsList>
           <>
-            {result.result.raw.cat_platform && (
-              <>
+            {result.result.raw.author && (
+              <div className="field">
                 <span className="field-label">
-                  <AtomicText value="Platform" />
+                  <AtomicText value="Author" />
                 </span>
-                <AtomicResultText field="cat_platform" />
-              </>
+                <AtomicResultText field="author" />
+              </div>
             )}
-            {result.result.raw.cat_condition && (
-              <>
+            {result.result.raw.source && (
+              <div className="field">
                 <span className="field-label">
-                  <AtomicText value="Condition" />
+                  <AtomicText value="Source" />
                 </span>
-                <AtomicResultText field="cat_condition" />
-              </>
+                <AtomicResultText field="source" />
+              </div>
             )}
-            {result.result.raw.cat_categories && (
-              <>
+            {result.result.raw.language && (
+              <div className="field">
                 <span className="field-label">
-                  <AtomicText value="Tags" />
+                  <AtomicText value="Language" />
                 </span>
-                <AtomicResultMultiValueText field="cat_categories" />
-              </>
+                <AtomicResultMultiValueText field="language" />
+              </div>
+            )}
+            {result.result.raw.filetype && (
+              <div className="field">
+                <span className="field-label">
+                  <AtomicText value="Filetype" />
+                </span>
+                <AtomicResultText field="filetype" />
+              </div>
             )}
           </>
         </AtomicResultFieldsList>